Address PR comments
- rename file with `unit` in the name to not have `unit` in the name
- Move collection.<method> tests from this PR into the collection.test.ts file
- Delete old skipped tests that test --no-deprecation in NodeJS
- Remove the `done` function from synchronous tests
